Improved timer for Windows
The current timer for Windows has two drawbacks: 1) on Windows 7 and
older, it is not very precise; 2) one must set NTDDI correctly (for me,
the current code would not directly compile on Windows 7). This proposed
timer uses QueryPerformanceCounter, which has the same precision as
GetSystemTimePreciseAsFileTime and works on all versions of Windows
since 2000.