Improved CPU timer for Windows
In the spirit of compact unit testing code, I'd like to propose a
compacted version of the CPU timing code for Windows. Use of
FileTimeToSystemTime() can be avoided as the GetProcessTimes() already
provides the time in units of 100 ns (so division by 10,000,000
suffices). Conversion of FILETIME to ULARGE_INTEGER type is implemented
as recommended by Microsoft. Additional bugfix: the error return code of
GetProcessTimes() is '0', not '-1'. Unfortunately, the resolution of
GetProcessTimes() is low, but Windows does not offer a better API so
it'll have to do.