Improved report for mu_assert_double_eq
Currently, the number of significant figures printed when mu_assert_double_eq fails is undefined. This leads to reports such as: "0.11915440 expected but was 0.11915440" whenever the accuracy in the assert is finer than the significance in printing. 

This code change causes the number of significant figures to vary with the value of MINUNIT_EPSILON. MINUNIT_EPSILON is redefined to be the additive inverse of the exponent. Conversion to 10^(-MINUNIT_EPSILON) is done by the preprocessor (avoiding the use of pow() ).

Examples of the new behavior: 
MINUNIT_EPSILON=4 -> "0.1192 expected but was 0.1194"
MINUNIT_EPSILON=12 -> "0.119154425113 expected but was 0.119354425113"