Improved report for mu_assert_double_eq (#18)
Currently, the number of significant figures printed when mu_assert_double_eq fails is undefined. This leads to reports such as: "0.11915440 expected but was 0.11915440" whenever the accuracy in the assert is finer than the significance in printing. 

This code change causes the number of significant figures to vary with the value of MINUNIT_EPSILON. The number of significant figures is calculated from MINUNIT_EPSILON by `minunit_significant_figures = 1 - log10(MINUNIT_EPSILON)`. The number is increased by one in order to print figures in cases where MINUNIT_EPSILON's significand is unequal to 1.

Examples of the new behavior: 
MINUNIT_EPSILON=1e-4 -> "0.1192 expected but was 0.1194"
MINUNIT_EPSILON=1e-12 -> "0.119154425113 expected but was 0.119354425113"