Fix clang-tidy warnings
With a strict clang-tidy setup it will warn about ignored return values
from fflush and snprintf. snprintf's return value will always be
zero-terminated, so it's fine to use even if it happens to be truncated.
For fflush(), we can't do anything with an error code.

By casting the result to void these are silenced.